Script changes to decouple from storage. Fix:  https://github.com/Azure/azure-webjobs-sdk-script/issues/1584
Depends on SDK with this PR: https://github.com/Azure/azure-webjobs-sdk/pull/1184
