+++ src/WebJobs.Script/WebJobs.Script.csproj
@@ -509,7 +509,7 @@
     <Compile Include="Extensions\IMetricsLoggerExtensions.cs" />
     <Compile Include="Extensions\TraceWriterExtensions.cs" />
     <Compile Include="GlobalSuppressions.cs" />
-    <Compile Include="Host\BlobLeaseManager.cs" />
+    <Compile Include="Host\PrimaryHostCoordinator.cs" />
     <Compile Include="Host\IScriptHostEnvironment.cs" />
     <Compile Include="Host\NullScriptHostEnvironment.cs" />
     <Compile Include="Host\ScriptHost.cs" />
