+++ test/WebJobs.Script.Tests.Integration/PrimaryHostCoordinatorTests.cs
@@ -17,8 +17,17 @@ using Xunit;
 
 namespace Microsoft.Azure.WebJobs.Script.Tests
 {
-    public class BlobLeaseManagerTests
+    public class PrimaryHostCoordinatorTests
     {
+        // Helper to get a real Blob Lease based IDistributedLockManager
+        private static IDistributedLockManager CreateLockManager()
+        {
+            JobHostConfiguration config = new JobHostConfiguration();
+            var host = new JobHost(config);
+            var lockManager = (IDistributedLockManager)host.Services.GetService(typeof(IDistributedLockManager));
+            return lockManager;
+        }
+
         [Fact]
         public async Task HasLease_WhenLeaseIsAcquired_ReturnsTrue()
         {
@@ -27,11 +36,9 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             string instanceId = Guid.NewGuid().ToString();
             var traceWriter = new TestTraceWriter(System.Diagnostics.TraceLevel.Verbose);
 
-            using (var manager = await BlobLeaseManager.CreateAsync(connectionString, TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null))
+            using (var manager = PrimaryHostCoordinator.Create(CreateLockManager(), TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null))
             {
                 await TestHelpers.Await(() => manager.HasLease);
-
-                Assert.Equal(instanceId, manager.LeaseId);
             }
 
             await ClearLeaseBlob(hostId);
@@ -51,11 +58,11 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             // Acquire a lease on the host lock blob
             string leaseId = await blob.AcquireLeaseAsync(TimeSpan.FromSeconds(15));
 
-            BlobLeaseManager manager = null;
+            PrimaryHostCoordinator manager = null;
 
             try
             {
-                manager = await BlobLeaseManager.CreateAsync(connectionString, TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null);
+                manager = PrimaryHostCoordinator.Create(CreateLockManager(), TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null);
                 manager.HasLeaseChanged += (s, a) => resetEvent.Set();
             }
             finally
@@ -65,12 +72,11 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
 
             resetEvent.Wait(TimeSpan.FromSeconds(15));
             bool hasLease = manager.HasLease;
-            string actualLeaseId = manager.LeaseId;
+
             manager.Dispose();
 
             Assert.True(resetEvent.IsSet);
-            Assert.True(hasLease, $"{nameof(BlobLeaseManager.HasLease)} was not correctly set to 'true' when lease was acquired.");
-            Assert.Equal(instanceId, actualLeaseId);
+            Assert.True(hasLease, $"{nameof(PrimaryHostCoordinator.HasLease)} was not correctly set to 'true' when lease was acquired.");
 
             await ClearLeaseBlob(hostId);
         }
@@ -86,10 +92,12 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             var traceWriter = new TestTraceWriter(TraceLevel.Verbose);
             var resetEvent = new ManualResetEventSlim();
 
-            BlobLeaseManager manager = null;
+            PrimaryHostCoordinator manager = null;
             string tempLeaseId = null;
 
-            using (manager = new BlobLeaseManager(blob, TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null, TimeSpan.FromSeconds(3)))
+            var lockManager = CreateLockManager();
+            var renewalInterval = TimeSpan.FromSeconds(3);
+            using (manager = PrimaryHostCoordinator.Create(lockManager, TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null, renewalInterval))
             {
                 try
                 {
@@ -98,7 +106,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
                     manager.HasLeaseChanged += (s, a) => resetEvent.Set();
 
                     // Release the manager's lease and acquire one with a different id
-                    await blob.ReleaseLeaseAsync(new AccessCondition { LeaseId = manager.LeaseId });
+                    await lockManager.ReleaseLockAsync(manager.LockHandle, CancellationToken.None);
                     tempLeaseId = await blob.AcquireLeaseAsync(TimeSpan.FromSeconds(30), Guid.NewGuid().ToString());
                 }
                 finally
@@ -113,77 +121,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             }
 
             Assert.True(resetEvent.IsSet);
-            Assert.False(manager.HasLease, $"{nameof(BlobLeaseManager.HasLease)} was not correctly set to 'false' when lease lost.");
-
-            await ClearLeaseBlob(hostId);
-        }
-
-        [Fact]
-        public void AcquiringLease_WithServerError_LogsAndRetries()
-        {
-            string hostId = Guid.NewGuid().ToString();
-            string instanceId = Guid.NewGuid().ToString();
-            var traceWriter = new TestTraceWriter(TraceLevel.Verbose);
-            var renewResetEvent = new ManualResetEventSlim();
-
-            var results = new Queue<Task<string>>();
-            results.Enqueue(Task.FromException<string>(new StorageException(new RequestResult { HttpStatusCode = 500 }, "test", null)));
-            results.Enqueue(Task.FromResult(hostId));
-
-            var blobMock = new Mock<ICloudBlob>();
-            blobMock.Setup(b => b.AcquireLeaseAsync(It.IsAny<TimeSpan>(), It.IsAny<string>()))
-                .Returns(() => results.Dequeue());
-
-            blobMock.Setup(b => b.RenewLeaseAsync(It.IsAny<AccessCondition>()))
-                .Returns(() => Task.Delay(1000))
-                .Callback(() => renewResetEvent.Set());
-
-            BlobLeaseManager manager;
-            using (manager = new BlobLeaseManager(blobMock.Object, TimeSpan.FromSeconds(5), hostId, instanceId, traceWriter, null))
-            {
-                renewResetEvent.Wait(TimeSpan.FromSeconds(10));
-            }
-
-            Assert.True(renewResetEvent.IsSet);
-            Assert.True(manager.HasLease);
-            Assert.True(traceWriter.Traces.Any(t => t.Message.Contains("Server error")));
-        }
-
-        [Fact]
-        public async Task Renew_WhenBlobIsDeleted_RecreatesBlob()
-        {
-            string hostId = Guid.NewGuid().ToString();
-            string instanceId = Guid.NewGuid().ToString();
-            var traceWriter = new TestTraceWriter(TraceLevel.Verbose);
-            var renewResetEvent = new ManualResetEventSlim();
-            string connectionString = AmbientConnectionStringProvider.Instance.GetConnectionString(ConnectionStringNames.Storage);
-
-            ICloudBlob blob = await GetLockBlobAsync(connectionString, hostId);
-
-            var blobMock = new Mock<ICloudBlob>();
-            blobMock.Setup(b => b.AcquireLeaseAsync(It.IsAny<TimeSpan>(), It.IsAny<string>()))
-                .Returns(() => Task.FromResult(hostId));
-
-            blobMock.Setup(b => b.RenewLeaseAsync(It.IsAny<AccessCondition>()))
-                .Returns(() => Task.FromException<string>(new StorageException(new RequestResult { HttpStatusCode = 404 }, "test", null)))
-                .Callback(() => Task.Delay(1000).ContinueWith(t => renewResetEvent.Set()));
-
-            blobMock.SetupGet(b => b.ServiceClient).Returns(blob.ServiceClient);
-
-            // Delete the blob
-            await blob.DeleteIfExistsAsync();
-
-            using (var manager = new BlobLeaseManager(blobMock.Object, TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null, TimeSpan.FromSeconds(3)))
-            {
-                renewResetEvent.Wait(TimeSpan.FromSeconds(10));
-
-                await TestHelpers.Await(() => manager.HasLease);
-            }
-
-            bool blobExists = await blob.ExistsAsync();
-
-            Assert.True(renewResetEvent.IsSet);
-            Assert.True(blobExists);
+            Assert.False(manager.HasLease, $"{nameof(PrimaryHostCoordinator.HasLease)} was not correctly set to 'false' when lease lost.");
 
             await ClearLeaseBlob(hostId);
         }
@@ -197,7 +135,8 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
 
             var traceWriter = new TestTraceWriter(System.Diagnostics.TraceLevel.Verbose);
 
-            using (var manager = await BlobLeaseManager.CreateAsync(connectionString, TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null))
+            var lockManager = CreateLockManager();
+            using (var manager = PrimaryHostCoordinator.Create(lockManager, TimeSpan.FromSeconds(15), hostId, instanceId, traceWriter, null))
             {
                 await TestHelpers.Await(() => manager.HasLease);
             }
@@ -229,11 +168,12 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             var traceWriter = new TestTraceWriter(TraceLevel.Verbose);
             var renewResetEvent = new ManualResetEventSlim();
 
-            var blobMock = new Mock<ICloudBlob>();
-            blobMock.Setup(b => b.AcquireLeaseAsync(It.IsAny<TimeSpan>(), It.IsAny<string>()))
-                .Returns(() => Task.FromResult(hostId));
+            var blobMock = new Mock<IDistributedLockManager>();
+            blobMock.Setup(b => b.TryLockAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(),
+                It.IsAny<TimeSpan>(), It.IsAny<CancellationToken>()))
+                .Returns(() => Task.FromResult<IDistributedLock>(new FakeLock()));
 
-            using (var manager = new BlobLeaseManager(blobMock.Object, TimeSpan.FromSeconds(5), hostId, instanceId, traceWriter, null))
+            using (var manager = new PrimaryHostCoordinator(blobMock.Object, TimeSpan.FromSeconds(5), hostId, instanceId, traceWriter, null))
             {
                 renewResetEvent.Wait(TimeSpan.FromSeconds(10));
 
@@ -254,15 +194,16 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             var traceWriter = new TestTraceWriter(TraceLevel.Verbose);
             var renewResetEvent = new ManualResetEventSlim();
 
-            var blobMock = new Mock<ICloudBlob>();
-            blobMock.Setup(b => b.AcquireLeaseAsync(It.IsAny<TimeSpan>(), It.IsAny<string>()))
-                .Returns(() => Task.FromResult(hostId));
+            var blobMock = new Mock<IDistributedLockManager>();
+            blobMock.Setup(b => b.TryLockAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(),
+                It.IsAny<TimeSpan>(), It.IsAny<CancellationToken>()))
+                .Returns(() => Task.FromResult<IDistributedLock>(new FakeLock()));
 
-            blobMock.Setup(b => b.RenewLeaseAsync(It.IsAny<AccessCondition>()))
-                .Returns(() => Task.FromException(new StorageException(new RequestResult { HttpStatusCode = 409 }, "test", null)))
+            blobMock.Setup(b => b.RenewAsync(It.IsAny<IDistributedLock>(), It.IsAny<CancellationToken>()))
+                .Returns(() => Task.FromException<bool>(new StorageException(new RequestResult { HttpStatusCode = 409 }, "test", null)))
                 .Callback(() => renewResetEvent.Set());
 
-            using (var manager = new BlobLeaseManager(blobMock.Object, TimeSpan.FromSeconds(5), hostId, instanceId, traceWriter, null))
+            using (var manager = new PrimaryHostCoordinator(blobMock.Object, TimeSpan.FromSeconds(5), hostId, instanceId, traceWriter, null))
             {
                 renewResetEvent.Wait(TimeSpan.FromSeconds(10));
                 await TestHelpers.Await(() => traceWriter.Traces.Count == 2, 5000, 500);
@@ -287,8 +228,9 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             string connectionString = AmbientConnectionStringProvider.Instance.GetConnectionString(ConnectionStringNames.Storage);
             var traceWriter = new TestTraceWriter(TraceLevel.Verbose);
 
-            using (var manager1 = await BlobLeaseManager.CreateAsync(connectionString, TimeSpan.FromSeconds(15), hostId1, instanceId, traceWriter, null))
-            using (var manager2 = await BlobLeaseManager.CreateAsync(connectionString, TimeSpan.FromSeconds(15), hostId2, instanceId, traceWriter, null))
+            var lockManager = CreateLockManager();
+            using (var manager1 = PrimaryHostCoordinator.Create(lockManager, TimeSpan.FromSeconds(15), hostId1, instanceId, traceWriter, null))
+            using (var manager2 = PrimaryHostCoordinator.Create(lockManager, TimeSpan.FromSeconds(15), hostId2, instanceId, traceWriter, null))
             {
                 Task manager1Check = TestHelpers.Await(() => manager1.HasLease);
                 Task manager2Check = TestHelpers.Await(() => manager2.HasLease);
@@ -304,10 +246,10 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             CloudStorageAccount account = CloudStorageAccount.Parse(accountConnectionString);
             CloudBlobClient client = account.CreateCloudBlobClient();
 
-            var container = client.GetContainerReference(BlobLeaseManager.HostContainerName);
+            var container = client.GetContainerReference(PrimaryHostCoordinator.HostContainerName);
 
             await container.CreateIfNotExistsAsync();
-            CloudBlockBlob blob = container.GetBlockBlobReference(BlobLeaseManager.GetBlobName(hostId));
+            CloudBlockBlob blob = container.GetBlockBlobReference(PrimaryHostCoordinator.GetBlobName(hostId));
             if (!await blob.ExistsAsync())
             {
                 await blob.UploadFromStreamAsync(new MemoryStream());
@@ -330,6 +272,13 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             }
 
             await blob.DeleteIfExistsAsync();
+        }
+
+        private class FakeLock : IDistributedLock
+        {
+            public string LockId => "lockid";
+
+            public Task LeaseLost => throw new NotImplementedException();
         }
     }
 }
