+++ test/WebJobs.Script.Tests/Binding/SendGridScriptBindingProviderTests.cs
@@ -6,6 +6,7 @@ using System.Diagnostics;
 using System.Reflection;
 using Microsoft.Azure.WebJobs.Extensions.SendGrid;
 using Microsoft.Azure.WebJobs.Host.Config;
+using Microsoft.Azure.WebJobs.Script;
 using Microsoft.Azure.WebJobs.Script.Binding;
 using Microsoft.Azure.WebJobs.Script.Extensibility;
 using Newtonsoft.Json.Linq;
@@ -25,7 +26,8 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             JObject hostMetadata = new JObject();
 
             var provider = new GeneralScriptBindingProvider(config, hostMetadata, traceWriter);
-            provider.CompleteInitialization();
+            var metadataProvider = new JobHost(config).CreateMetadataProvider();
+            provider.CompleteInitialization(metadataProvider);
             _provider = provider;
         }
